Fixed Broken Maze Issue
Minor Issue in the Maze generation script fixed.

When neighbours are checked to see if they have a parent (i.e. - already added to the maze), it was possible to select the root as the root does not have a parent.

A string was added to the root so it was not null.